Inicio da lista base
