+++ .vscode/settings.json
@@ -1,7 +1,8 @@
 {
     "files.associations": {
         "stdlib.h": "c",
-        "stdio.h": "c"
+        "stdio.h": "c",
+        "cstdlib": "cpp"
     },
     "C_Cpp.errorSquiggles": "Enabled"
 }
