--- Estruturas/Linked_List.cpp
@@ -1,21 +1,0 @@
-# include "No.hpp"
-
-template <typename type>
-class Lista_Encadeada
-{
-protected:
-    int tamanho;
-    
-public:
-    Lista_Encadeada(/* args */);
-    ~Lista_Encadeada();
-};
-
-Lista_Encadeada::Lista_Encadeada(/* args */)
-{
-}
-
-Lista_Encadeada::~Lista_Encadeada()
-{
-}
-
